changed num1 to int

+++ stats.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-  double num1;
+  int num1;
   double num2;
   std::cout << "Please enter the two numbers: ";
   std::cin >> num1 >> num2;
